Move RedisClient from package cache to utils
